api.c: pull out TLS 1.3 specific tests
